made changes in company, intern, person, social medias & field experiences services to fix for bugs on admin dashboard when updating or deleting records

+++ src/interns/interns.service.ts
@@ -1,29 +1,45 @@
-import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
-import { CreateInternInput } from './dto/create-intern.input';
-import { UpdateInternInput } from './dto/update-intern.input';
-import { Intern } from './entities/intern.entity';
+import { Injectable } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { Repository } from "typeorm";
+import { CreateInternInput } from "./dto/create-intern.input";
+import { UpdateInternInput } from "./dto/update-intern.input";
+import { Intern } from "./entities/intern.entity";
 
 @Injectable()
 export class InternsService {
   constructor(
     @InjectRepository(Intern)
-    private readonly internRepository: Repository<Intern>,
+    private readonly internRepository: Repository<Intern>
   ) {}
 
-  create(createInternInput: CreateInternInput): Promise<Intern> {
-    const newIntern = this.internRepository.create(createInternInput);
-    return this.internRepository.save(newIntern);
+  async create(createInternInput: CreateInternInput): Promise<Intern> {
+    const newIntern = await this.internRepository.create(createInternInput);
+    return await this.internRepository.save(newIntern);
   }
 
   findAll(): Promise<Intern[]> {
     return this.internRepository.find();
   }
 
+  findAllByCompanyId(companyId: number): Promise<Intern[]> {
+    return this.internRepository.find({
+      where: {
+        companyId,
+      },
+    });
+  }
+
+  findByStudentId(studentId: number): Promise<Intern[]> {
+    return this.internRepository.find({
+      where: {
+        studentId,
+      },
+    });
+  }
+
   findOneById(id: number): Promise<Intern> {
     return this.internRepository.findOneOrFail(id, {
-      relations: ['student'],
+      relations: ["student"],
     });
   }
 
@@ -36,6 +52,8 @@ export class InternsService {
 
   async remove(id: number): Promise<Intern> {
     const intern = await this.findOneById(id);
-    return this.internRepository.remove(intern);
+    if (intern) {
+      return await this.internRepository.remove(intern);
+    } else return;
   }
 }
